requirements.txt
fastapi
uvicorn
psycopg2-binary
httpx
pytest
pydantic
python-dotenv
redis
pydantic[email]